Refactor platform structure and enhance frontend experience
- Updated contributing guidelines to reflect changes in frontend technology from Angular to static HTML/CSS/JavaScript.
- Modified deployment documentation to include new frontend service and updated port configurations.
- Enhanced local setup instructions to clarify frontend access and Docker usage.
- Introduced a new static frontend served by Nginx, including a polished single-page UI for analysis submission and results viewing.
- Updated backend configurations to support new frontend and improved CORS settings.
- Added health checks and improved Docker Compose configurations for better service management.
- Implemented various code improvements in the AI worker and backend, including type casting and serialization enhancements.
- Added unit tests for AWS configuration to ensure proper serialization of LocalDateTime.
- Introduced a new logo asset for branding.

+++ ai-worker/models/sequence_embedder.py
@@ -50,4 +50,4 @@ def compute_mutation_distance(wild_type: str, mutant: str, k: int = 6) -> float:
     norm = np.linalg.norm(vec_wt) * np.linalg.norm(vec_mut)
 
     cosine_similarity = dot / norm if norm > 0 else 0.0
-    return round(1.0 - cosine_similarity, 6)
+    return float(round(1.0 - cosine_similarity, 6))
